--- etc.singularity.js
@@ -1,377 +1,0 @@
-/** @param {NS} ns **/
-
-function scan(ns, parent, server, list) {
-	const children = ns.scan(server);
-	for (let child of children) {
-		if (parent == child) {
-			continue;
-		}
-		list.push(child);
-
-		scan(ns, server, child, list);
-	}
-}
-
-export function list_servers(ns) {
-	const list = [];
-	scan(ns, '', 'home', list);
-	return list;
-}
-
-const spiralize_matrix = (matrix) => {
-	var result = [];
-	while (matrix.length > 0) {
-		result.push(...matrix.shift());
-		try {
-			matrix = matrix[0].map((val, idx) => matrix.map(row => row[idx]).reverse()); // i only know how to rotate clockwise
-			matrix = matrix[0].map((val, idx) => matrix.map(row => row[idx]).reverse()); // this solution wants it counterclockwise
-			matrix = matrix[0].map((val, idx) => matrix.map(row => row[idx]).reverse()); // but three clockwise rotations is counter clockwise!
-		} catch (e) {
-			result.push(...matrix);
-		}
-	}
-	return result;
-};
-
-function twts(N, K)
-{   
-    let dp = Array.from({length: N +1}, (_, i) => 0);
-    dp[0] = 1;
-    for(let row = 1; row < K + 1; row++)
-    {
-        for(let col = 1; col < N + 1; col++)
-        {
-            if (col >= row)
-                dp[col] = dp[col] + dp[col - row];
-          }
-    }
- 
-    return(dp[N]-1);
-}
-
-
-/**
- * Dynamic Programming solution.
- * Complexity: O(n)
- *
- * @param {Number[]} inputArray
- * @return {Number[]}
- */
-function dpMaximumSubarray(inputArray) {
-	// We iterate through the inputArray once, using a greedy approach to keep track of the maximum
-	// sum we've seen so far and the current sum.
-	//
-	// The currentSum variable gets reset to 0 every time it drops below 0.
-	//
-	// The maxSum variable is set to -Infinity so that if all numbers are negative, the highest
-	// negative number will constitute the maximum subarray.
-
-	let maxSum = -Infinity;
-	let currentSum = 0;
-
-	// We need to keep track of the starting and ending indices that contributed to our maxSum
-	// so that we can return the actual subarray. From the beginning let's assume that whole array
-	// is contributing to maxSum.
-	let maxStartIndex = 0;
-	let maxEndIndex = inputArray.length - 1;
-	let currentStartIndex = 0;
-
-	inputArray.forEach((currentNumber, currentIndex) => {
-		currentSum += currentNumber;
-
-		// Update maxSum and the corresponding indices if we have found a new max.
-		if (maxSum < currentSum) {
-			maxSum = currentSum;
-			maxStartIndex = currentStartIndex;
-			maxEndIndex = currentIndex;
-		}
-
-		// Reset currentSum and currentStartIndex if currentSum drops below 0.
-		if (currentSum < 0) {
-			currentSum = 0;
-			currentStartIndex = currentIndex + 1;
-		}
-	});
-
-	let max = inputArray.slice(maxStartIndex, maxEndIndex + 1);
-	return max.reduce((a, b) => a + b, 0);
-}
-
-
-const find_best_single_trade = (array) => {
-	let best_single_trade = 0;
-	let idx = 0;
-	for (let cur of array) {
-		for (let rest of array.slice(idx)) {
-			if ((rest - cur) > best_single_trade) {
-				best_single_trade = rest - cur;
-			}
-		}
-		idx++;
-	}
-	return best_single_trade;
-};
-
-const find_cumulative_profit = (array) => {
-    return array.reduce(function (accumulator, current, idx, arr) {
-        if (arr[idx+1] > current){
-            // console.log("Index ", idx, " profit ", (arr[idx+1]-current), " cumulatively ", accumulator);
-            return accumulator + (arr[idx+1]-current);
-        } else { return accumulator || 0; }
-    }, 0);
-};
-
-const unique_paths_with_obstacles = (arr) => {
-	let r = arr.length,
-		c = arr[0].length;
-
-	// create a 2D-matrix and initializing
-	// with value 0
-	let paths = new Array(r);
-	for (let i = 0; i < r; i++) {
-		paths[i] = new Array(c);
-		for (let j = 0; j < c; j++) {
-			paths[i][j] = 0;
-		}
-	}
-
-	// Initializing the left corner if
-	// no obstacle there
-	if (arr[0][0] == 0)
-		paths[0][0] = 1;
-
-	// Initializing first column of
-	// the 2D matrix
-	for (let i = 1; i < r; i++) {
-		// If not obstacle
-		if (arr[i][0] == 0)
-			paths[i][0] = paths[i - 1][0];
-	}
-
-	// Initializing first row of the 2D matrix
-	for (let j = 1; j < c; j++) {
-
-		// If not obstacle
-		if (arr[0][j] == 0)
-			paths[0][j] = paths[0][j - 1];
-	}
-
-	for (let i = 1; i < r; i++) {
-		for (let j = 1; j < c; j++) {
-
-			// If current cell is not obstacle
-			if (arr[i][j] == 0)
-				paths[i][j] = paths[i - 1][j] +
-				paths[i][j - 1];
-		}
-	}
-
-	// Returning the corner value
-	// of the matrix
-	return paths[r - 1];
-};
-
-
-function minSumPath(arr) {
-
-	// For storing the result
-	// in a 1-D array, and
-	// simultaneously updating
-	// the result.
-	let memo = [];
-	let n = arr.length - 1;
-
-	// For the bottom row
-	for (let i = 0; i < arr[n].length; i++)
-		memo[i] = arr[n][i];
-
-	// Calculation of the
-	// remaining rows, in
-	// bottom up manner.
-	for (let i = arr.length - 2; i >= 0; i--)
-		for (let j = 0; j < arr[i].length; j++)
-			memo[j] = arr[i][j] +
-			Math.min(memo[j],
-				memo[j + 1]);
-
-	// Return the
-	// top element
-	return memo[0];
-}
-
-
-function range(size, startAt = 0) {
-    return [...Array(size).keys()].map(i => i + startAt);
-}
-
-const array_jumping_generate_graph = (solution_array) => {
-    let graph = new Map();
-    for (let i = 0; i < solution_array.length; i++) {
-        graph.set(i, range(solution_array[i], i+1)
-        );
-    }
-    return graph;
-};
-
-const array_jumping_traverse_graph = (graph, source) => {
-    const stack = [ source ];
-    const result = [];
-    const visited = {};
-    visited [ source ] = true;
-    let current;
-
-    while (stack.length > 0) {
-        current = stack.pop();
-        // console.log(current);
-        if (current >= graph.size) { return true; }
-        result.push(current);
-        try {
-            graph.get(current).forEach(neighbor => {
-                if (!visited[neighbor]) {
-                    visited[neighbor] = true;
-                    stack.push(neighbor);
-                }
-            });
-        } catch (e) {
-            console.log(e);
-            return true;
-        }
-    }
-    return false;
-};
-    
-
-const array_jumping_can_win = (solution_array) => {
-    let graph = array_jumping_generate_graph(solution_array);
-    return array_jumping_traverse_graph(graph, 0);
-};
-
-const find_largest_prime_factor = (target) => {
-    var x = 2;
-    while (x <= target) {
-        if (target % x == 0) {
-            target /= x;
-        } else {
-            x++;
-        }
-    }
-    return x;
-};
-
-export async function main(ns) {
-	/*****************
-	solve coding contracts
-	*****************/
-	let servers = list_servers(ns);
-
-	while (true) {
-
-		for (let server of servers) {
-			if (ns.ls(server, ".cct").length > 0) {
-				var contractName = ns.ls(server, ".cct")[0];
-				var result;
-				switch (ns.codingcontract.getContractType(contractName, server)) {
-					case "Spiralize Matrix":
-						result = spiralize_matrix(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Subarray with Maximum Sum":
-						result = dpMaximumSubarray(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Algorithmic Stock Trader I":
-						result = find_best_single_trade(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Algorithmic Stock Trader II":
-						result = find_cumulative_profit(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Unique Paths in a Grid II":
-						result = unique_paths_with_obstacles(ns.codingcontract.getData(contractName, server))[unique_paths_with_obstacles(ns.codingcontract.getData(contractName, server)).length - 1];
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Minimum Path Sum in a Triangle":
-						result = minSumPath(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Array Jumping Game":
-						result = array_jumping_can_win(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Find Largest Prime Factor":
-						result = find_largest_prime_factor(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					case "Total Ways to Sum":
-						result = twts(ns.codingcontract.getData(contractName, server));
-						ns.tprint(ns.codingcontract.attempt(result, contractName, server, {returnReward: true}));
-						break;
-					default:
-						break;
-				}
-
-			}
-		}
-
-	/*****************
-	Join factions
-	*****************/
-
-		let faction_list = [
-			"Netburners",
-			"CyberSec",
-			"NiteSec",
-			"BitRunners",
-			"Daedalus",
-			"Sector-12",
-			"Tian Di Hui",
-			"The Black Hand"
-		];
-		for (let faction of faction_list) {
-			try {
-				ns.joinFaction(faction);
-			} catch (e) {}
-		}
-		if (ns.getFactionRep("BitRunners") < 250000) {
-			ns.workForFaction("BitRunners", "Hacking Contracts");
-		 }
-		 if (ns.getFactionRep("NiteSec") < 112500) {
-			ns.workForFaction("NiteSec", "Hacking Contracts");
-		 }
-
-		 if (ns.getFactionRep("CyberSec") < 12500) {
-			ns.workForFaction("CyberSec", "Hacking Contracts");
-		 }
-		 
-	/*****************
-	Get software
-	*****************/
-		try {
-			ns.purchaseTor();
-		} catch (e) {}
-
-		let software_list = [
-			"BruteSSH.exe",
-			"FTPCrack.exe",
-			"relaySMTP.exe",
-			"HTTPWorm.exe",
-			"SQLInject.exe"
-		];
-
-		for (let software of software_list) {
-			try {
-				ns.purchaseProgram(software);
-			} catch(e) {}
-		}
-	/*****************
-	Upgrade home
-	*****************/
-
-		try {
-			ns.upgradeHomeRam();
-		} catch (e) {}
-
-		await ns.sleep(60000);
-	}
-}
