+++ lib.gamestates.so.js
@@ -146,21 +146,19 @@ export class gsNewGame extends DefaultGameStage {
 
         while (servers.filter(s => s.admin).length < 3) {
             ns.exec("sbin.sudo.js", "home", 1, "n00dles");
-            await ns.sleep(400);
+            await ns.sleep(150);
             ns.exec("sbin.sudo.js", "home", 1, "foodnstuff");
-            await ns.sleep(400);
+            await ns.sleep(800);
             ns.exec("sbin.scp.js", "home", 1, "n00dles", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
+            await ns.sleep(150);
             ns.exec("sbin.scp.js", "home", 1, "foodnstuff", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
+            await ns.sleep(150);
         }
 
         ns.exec("bin.gr.loop.js", "n00dles", 1, "foodnstuff");
-        await ns.sleep(100);
         ns.exec("bin.wk.loop.js", "n00dles", 1, "foodnstuff");
-        await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "foodnstuff", 9, "foodnstuff");
-        await ns.sleep(100);
+        await ns.sleep(800);
     
         while (player.level < 5) {
             let curtime = new Date().valueOf();
@@ -177,11 +175,9 @@ export class gsNewGame extends DefaultGameStage {
         }
         
         ns.exec("sbin.scp.js", "home", 1, "sigma-cosmetics", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-        await ns.sleep(400);
+        await ns.sleep(800);
         ns.exec("bin.wk.loop.js", "sigma-cosmetics", 1, "foodnstuff");
-        await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "sigma-cosmetics", 8, "foodnstuff");
-        await ns.sleep(100);
 
         while (player.level < 10) {
             await ns.sleep(foodnstuff.hackTime + 55);
@@ -192,14 +188,11 @@ export class gsNewGame extends DefaultGameStage {
         while (servers.filter(s => s.admin).length < 5) {
             ns.exec("sbin.sudo.js", "home", 1, "joesguns");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "joesguns", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
-
+        ns.exec("sbin.scp.js", "home", 1, "joesguns", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
         ns.exec("bin.wk.loop.js", "joesguns", 1, "foodnstuff");
-        await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "joesguns", 8, "foodnstuff");
-        await ns.sleep(100);
 
         while (player.level < 20) {
             await ns.sleep(foodnstuff.hackTime + 55);
@@ -209,14 +202,12 @@ export class gsNewGame extends DefaultGameStage {
         while (servers.filter(s => s.admin).length < 6) {
             ns.exec("sbin.sudo.js", "home", 1, "nectar-net");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "nectar-net", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
+        ns.exec("sbin.scp.js", "home", 1, "nectar-net", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
 
         ns.exec("bin.wk.loop.js", "nectar-net", 1, "foodnstuff");
-        await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "nectar-net", 8, "foodnstuff");
-        await ns.sleep(100);
 
         while (player.level < 30) {
             await ns.sleep(foodnstuff.hackTime + 55);
@@ -227,10 +218,10 @@ export class gsNewGame extends DefaultGameStage {
         while (servers.filter(s => s.admin).length < 7) {
             ns.exec("sbin.sudo.js", "home", 1, "hong-fang-tea");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "hong-fang-tea", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
-        
+        ns.exec("sbin.scp.js", "home", 1, "hong-fang-tea", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
+    
         ns.exec("bin.wk.loop.js", "hong-fang-tea", 1, "foodnstuff");
         await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "hong-fang-tea", 8, "foodnstuff");
@@ -245,10 +236,10 @@ export class gsNewGame extends DefaultGameStage {
         while (servers.filter(s => s.admin).length < 8) {
             ns.exec("sbin.sudo.js", "home", 1, "harakiri-sushi");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "harakiri-sushi", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
 
+        ns.exec("sbin.scp.js", "home", 1, "harakiri-sushi", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
         ns.exec("bin.wk.loop.js", "harakiri-sushi", 1, "foodnstuff");
         await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "harakiri-sushi", 8, "foodnstuff");
@@ -260,12 +251,16 @@ export class gsNewGame extends DefaultGameStage {
             ns.exec("bin.hk.once.js", "home", home.threadCount(1.7), "foodnstuff");
         }
         while (servers.filter(s => s.admin).length < 9) {
+            ns.purchaseTor();
+            ns.purchaseProgram("brutessh.exe");
+
             ns.exec("sbin.sudo.js", "home", 1, "neo-net");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "neo-net", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
+
         
+        ns.exec("sbin.scp.js", "home", 1, "neo-net", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
         ns.exec("bin.wk.loop.js", "neo-net", 2, "foodnstuff");
         await ns.sleep(100);
         ns.exec("bin.gr.loop.js", "neo-net", 16, "foodnstuff");
@@ -281,10 +276,11 @@ export class gsNewGame extends DefaultGameStage {
         while (servers.filter(s => s.admin).length < 10) {
             ns.exec("sbin.sudo.js", "home", 1, "CSEC");
             await ns.sleep(400);
-            ns.exec("sbin.scp.js", "home", 1, "CSEC", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
-            await ns.sleep(400);
         }
-        
+
+        ns.exec("sbin.scp.js", "home", 1, "CSEC", "bin.hk.loop.js", "bin.wk.loop.js", "bin.gr.loop.js");
+        await ns.sleep(400);
+  
         ns.exec("bin.wk.loop.js", "CSEC", 4, "foodnstuff");
         await ns.sleep(100);
 
@@ -308,10 +304,11 @@ export class gsNewGame extends DefaultGameStage {
 
         await ns.sleep(1000);
         if (ns.ps("home").every(process => process.filename != "sbin.hnet.js")) {
-            ns.exec("sbin.hnet.js", "home");
+            let hnet_pid = ns.exec("sbin.hnet.js", "home");
         }
         
         await ns.sleep(5000);
+        ns.kill(hnet_pid);
         return {player,servers};
     }
